refactor(annotationTool): renamed annotationTool to annotation-tool
